Move DeleteLik to new home
